% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_shiny_helpers.R
\name{setup_cloudflux_python}
\alias{setup_cloudflux_python}
\title{Initial Setup for CloudFlux Python Environment}
\usage{
setup_cloudflux_python()
}
\description{
This function creates the necessary Miniconda environment and installs
the hybrid GPU/CPU dependencies.
}
